Split ethtool to dedicated module
